Refactored Project Structure
