+++ src/app/client/assets/fonts/Roboto-BoldItalic.ttf
