+++ src/app/server/models/quote.js
