+++ src/app/server/routes/static.js
